+++ yt/ytlib/table_client/chunk_writer_base.cpp
@@ -73,7 +73,7 @@ void TChunkWriterBase::CheckBufferCapacity()
 
 void TChunkWriterBase::FinalizeWriter()
 {
-    Meta.set_type(EChunkType::Table);
+    Meta.set_type(static_cast<int>(EChunkType::Table));
     Meta.set_version(FormatVersion);
 
     SetProtoExtension(Meta.mutable_extensions(), ChannelsExt);
@@ -82,7 +82,7 @@ void TChunkWriterBase::FinalizeWriter()
         MiscExt.set_uncompressed_data_size(EncodingWriter->GetUncompressedSize());
         MiscExt.set_compressed_data_size(EncodingWriter->GetCompressedSize());
         MiscExt.set_meta_size(Meta.ByteSize());
-        MiscExt.set_compression_codec(Options->CompressionCodec);
+        MiscExt.set_compression_codec(static_cast<int>(Options->CompressionCodec));
         MiscExt.set_data_weight(DataWeight);
         MiscExt.set_row_count(RowCount);
         MiscExt.set_value_count(ValueCount);
